Covid 19 Virtual Assistant
This project is inspired from the current situation of COVID-19. It mainly aims at providing assistance to people in times of crises just as a doctor gives. Program will ask the patient few simple questions and on the basis of mnemonics for Corona Virus Diagnosis (as shown below) outputs the result and notifies the user what step they should take next. Moreover, it can act as a management system by tracking the corona cases which involves adding and removing cases.