Tower of Hanoi
Tower of Hanoi is a mathematical puzzle invented by a French mathematician, “Edouard Lucas” in 1883. The puzzle consists of three rods and a number of disks of different sizes, which can slide onto any rod. The puzzle starts with the disks in a neat stack in ascending order of size on one rod, the smallest at the top, thus making a conical shape. The objective of the puzzle is to move the entire stack to another rod, obeying the following simple rules:
1.	Only one disk can be moved at a time.
2.	Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack or on an empty rod.
3.	No larger disk may be placed on top of a smaller disk.
	MINIMUM STEPS FORMULA:
                                          The minimal number of moves required to solve a Tower of Hanoi puzzle is 2n − 1, where n is the number of disks.

  
    Fig. “The Puzzle: Tower of Hanoi”
 
	OBJECTIVE:
                        To give the minimum steps in order to solve the famous mathematical puzzle “Tower of Hanoi”.

	FUNCTIONALITY:
The functionality of the source code of the program will mainly be based on the principle of resurrection. 
	The whole outcome mainly relies on the recursive function. 
	Input/output functions declaring header file.
	Formatted input/output writing functions declaring the header file.
	Mathematical functions declaring libraries.
	Figlet (generated using Ascii Figlet Generator).
	Console color changing command.
 
	USE-CASE MODEL:
The user will be prompt to enter number of disks onto the screen and the program will execute, giving steps of solution to shift the disks in the chronological order to the designated rod.

	SCOPE:
                                     It can be used by mathematicians and engineers to solve not only the above-mentioned puzzle but many similar real-world problems. 

	IMPLEMENTATION CONSTRAINTS:
                                                                   The IDE that will be used to build the program is Dev C++, predicting this will improve long-term supportability in addition to ease of development. 
	INTERFACE:
                     Windows.
	NOTEWORTHY HARDWARE SPECIFICATIONS:
	Processor – i3
	Hard Disk – 5 GB
	Memory – 1GB RAM