Vehicle Parking Management System
The main objective of this C++ Project on Vehicle Parking Management System is to manage the details of parking fees, customers, parking slots, types, government tax, employee and manager wages and vehicles entering and leaving the parking slots. This project will be totally built for administrative end and thus only the administrator is guaranteed the access. The purpose of the project is to build an application program to reduce the manual work for managing the Duration, Parking fees, Vehicles, Customers. It tracks all the details about the customers, parking slots and types of vehicles.